Move file source container and loader to separate module
