+++ src/databricks/labs/ucx/source_code/linters/files.py
@@ -1,13 +1,13 @@
 from __future__ import annotations  # for type hints
 
-import dataclasses
 import logging
 from collections.abc import Iterable, Callable
 from pathlib import Path
 import sys
 from typing import TextIO
 
-from databricks.labs.ucx.source_code.base import LocatedAdvice, CurrentSessionState, file_language, is_a_notebook
+from databricks.labs.ucx.source_code.base import LocatedAdvice, CurrentSessionState, is_a_notebook
+from databricks.labs.ucx.source_code.files import FileLoader
 from databricks.labs.ucx.source_code.python.python_ast import Tree
 from databricks.labs.ucx.source_code.notebooks.loaders import NotebookLoader
 from databricks.labs.ucx.source_code.notebooks.sources import FileLinter
@@ -17,7 +17,6 @@ from databricks.sdk.service.workspace import Language
 from databricks.labs.blueprint.tui import Prompts
 
 from databricks.labs.ucx.source_code.linters.context import LinterContext
-from databricks.labs.ucx.source_code.notebooks.cells import CellLanguage, PythonCodeAnalyzer
 from databricks.labs.ucx.source_code.graph import (
     BaseImportResolver,
     BaseFileResolver,
@@ -28,54 +27,10 @@ from databricks.labs.ucx.source_code.graph import (
     MaybeDependency,
     SourceContainer,
     DependencyResolver,
-    InheritedContext,
     DependencyGraphWalker,
 )
 
 logger = logging.getLogger(__name__)
-
-
-class LocalFile(SourceContainer):
-
-    def __init__(self, path: Path, source: str, language: Language):
-        self._path = path
-        self._original_code = source
-        # using CellLanguage so we can reuse the facilities it provides
-        self._language = CellLanguage.of_language(language)
-
-    @property
-    def path(self) -> Path:
-        return self._path
-
-    def build_dependency_graph(self, parent: DependencyGraph) -> list[DependencyProblem]:
-        if self._language is CellLanguage.PYTHON:
-            context = parent.new_dependency_graph_context()
-            analyzer = PythonCodeAnalyzer(context, self._original_code)
-            problems = analyzer.build_graph()
-            for idx, problem in enumerate(problems):
-                if problem.has_missing_path():
-                    problems[idx] = dataclasses.replace(problem, source_path=self._path)
-            return problems
-        # supported language that does not generate dependencies
-        if self._language is CellLanguage.SQL:
-            return []
-        logger.warning(f"Unsupported language: {self._language.language}")
-        return []
-
-    def build_inherited_context(self, graph: DependencyGraph, child_path: Path) -> InheritedContext:
-        if self._language is CellLanguage.PYTHON:
-            context = graph.new_dependency_graph_context()
-            analyzer = PythonCodeAnalyzer(context, self._original_code)
-            inherited = analyzer.build_inherited_context(child_path)
-            problems = list(inherited.problems)
-            for idx, problem in enumerate(problems):
-                if problem.has_missing_path():
-                    problems[idx] = dataclasses.replace(problem, source_path=self._path)
-            return dataclasses.replace(inherited, problems=problems)
-        return InheritedContext(None, False, [])
-
-    def __repr__(self):
-        return f"<LocalFile {self._path}>"
 
 
 class Folder(SourceContainer):
@@ -244,39 +199,6 @@ class LocalFileMigrator:
                 logger.info(f"Overwriting {path}")
                 f.write(code)
                 return True
-
-
-class StubContainer(SourceContainer):
-
-    def __init__(self, path: Path):
-        super().__init__()
-        self._path = path
-
-    def build_dependency_graph(self, parent: DependencyGraph) -> list[DependencyProblem]:
-        return []
-
-
-class FileLoader(DependencyLoader):
-    def load_dependency(self, path_lookup: PathLookup, dependency: Dependency) -> SourceContainer | None:
-        absolute_path = path_lookup.resolve(dependency.path)
-        if not absolute_path:
-            return None
-        language = file_language(absolute_path)
-        if not language:
-            return StubContainer(absolute_path)
-        for encoding in ("utf-8", "ascii"):
-            try:
-                code = absolute_path.read_text(encoding)
-                return LocalFile(absolute_path, code, language)
-            except UnicodeDecodeError:
-                pass
-        return None
-
-    def exists(self, path: Path) -> bool:
-        return path.exists()
-
-    def __repr__(self):
-        return "FileLoader()"
 
 
 class FolderLoader(FileLoader):
