+++ tests/unit/source_code/test_graph.py
@@ -4,7 +4,8 @@ from pathlib import Path
 import pytest
 
 from databricks.labs.ucx.source_code.base import Advisory, CurrentSessionState, LocatedAdvice
-from databricks.labs.ucx.source_code.linters.files import FileLoader, FolderLoader
+from databricks.labs.ucx.source_code.linters.files import FolderLoader
+from databricks.labs.ucx.source_code.files import FileLoader
 from databricks.labs.ucx.source_code.graph import (
     Dependency,
     DependencyGraph,
